+++ .gitignore
@@ -7,3 +7,4 @@
 .idea/
 .clangd
 .cache/
+.history/
